Implement low and high priority actor spawning queues.
DelegateChain holds a list of functors working as a chain of
responsibility, the first functor that knows how to deal with the
given input will return an indication of which chain (low or high)
the new item should be inserted into, and at which position.
This way gameplay code can inject its own logic for affecting the
order in which actors are spawned.
For convenience, StdArray wraps a TArray into a more standard
interface.